requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
better-profanity==0.7.0
build==0.3.0
certifi==2020.12.5
chardet==4.0.0
comtypes==1.1.8
cycler==0.10.0
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-multiselectfield==0.1.12
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
docopt==0.6.2
filelock==3.0.12
google==3.0.0
gunicorn==20.1.0
idna==2.10
Js2Py==0.70
kiwisolver==1.3.1
ldap3==2.9
matplotlib==3.3.4
mysqlclient==2.0.3
numpy==1.20.1
packaging==20.9
pep517==0.9.1
phonenumbers==8.12.22
Pillow==8.1.0
pipwin==0.5.1
psycopg2==2.8.6
pyasn1==0.4.8
PyAudio @ file:///C:/Users/Dell/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pydirectory==0.1.15
pyjsparser==2.7.1
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2021.1
pywin32==300
requests==2.25.1
six==1.15.0
soupsieve==2.2
SpeechRecognition==3.8.1
sqlparse==0.4.1
static3==0.7.0
tabulate==0.8.7
toml==0.10.2
tzlocal==2.1
urllib3==1.26.3
virtualenv==20.4.2
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
